More refactoring to allow removing TrackingPoint
