+++ semanticcpg/src/main/scala/io/shiftleft/semanticcpg/language/nodemethods/TrackingPointMethods.scala
@@ -1,24 +1,7 @@
 package io.shiftleft.semanticcpg.language.nodemethods
 
-import io.shiftleft.codepropertygraph.generated.{Operators, nodes}
-import io.shiftleft.semanticcpg.accesspath.{
-  AccessElement,
-  AccessPath,
-  AddressOf,
-  ConstantAccess,
-  Elements,
-  IndirectionAccess,
-  PointerShift,
-  TrackedBase,
-  TrackedFormalReturn,
-  TrackedLiteral,
-  TrackedMethodOrTypeRef,
-  TrackedNamedVariable,
-  TrackedReturnValue,
-  TrackedUnknown,
-  VariableAccess,
-  VariablePointerShift
-}
+import io.shiftleft.codepropertygraph.generated.nodes
+import io.shiftleft.semanticcpg.accesspath.{AccessElement, AccessPath, Elements, TrackedBase, TrackedReturnValue}
 import io.shiftleft.semanticcpg.language._
 import io.shiftleft.semanticcpg.utils.MemberAccess
 import org.slf4j.LoggerFactory
@@ -37,8 +20,6 @@ trait FakeTrackingPoint {
 
 // Utility methods for dealing with tracking points
 object TrackingPointMethodsBase {
-  private val logger = LoggerFactory.getLogger(getClass)
-  private var hasWarnedDeprecations = false
 
   //we don't want to expose this API everywhere, only when explicitly imported
   implicit class ImplicitsAPI(val node: nodes.TrackingPoint) extends AnyVal {
@@ -107,118 +88,16 @@ object TrackingPointMethodsBase {
   def toTrackedBase(node: nodes.TrackingPoint): TrackedBase = toTrackedBaseAndAccessPath(node)._1
   def toTrackedAccessPath(node: nodes.TrackingPoint): AccessPath = toTrackedBaseAndAccessPath(node)._2
 
-  private def toTrackedBaseAndAccessPathInternal(node: nodes.TrackingPoint): (TrackedBase, List[AccessElement]) =
+  private def toTrackedBaseAndAccessPathInternal(node: nodes.TrackingPoint): (TrackedBase, List[AccessElement]) = {
+
     node match {
       case node: nodes.ImplicitCall => (TrackedReturnValue(node), Nil)
       case node: nodes.PostExecutionCall =>
         toTrackedBaseAndAccessPathInternal(node._refOut.next().asInstanceOf[nodes.TrackingPoint])
-
-      case node: nodes.MethodParameterIn  => (TrackedNamedVariable(node.name), Nil)
-      case node: nodes.MethodParameterOut => (TrackedNamedVariable(node.name), Nil)
-
-      case node: nodes.Identifier    => (TrackedNamedVariable(node.name), Nil)
-      case node: nodes.Literal       => (TrackedLiteral(node), Nil)
-      case node: nodes.MethodRef     => (TrackedMethodOrTypeRef(node), Nil)
-      case node: nodes.TypeRef       => (TrackedMethodOrTypeRef(node), Nil)
-      case _: nodes.Return           => (TrackedFormalReturn, Nil)
-      case _: nodes.MethodReturn     => (TrackedFormalReturn, Nil)
-      case _: nodes.Unknown          => (TrackedUnknown, Nil)
-      case _: nodes.ControlStructure => (TrackedUnknown, Nil)
-      // FieldIdentifiers are only fake arguments, hence should not be tracked
-      case _: nodes.FieldIdentifier => (TrackedUnknown, Nil)
-      case block: nodes.Block =>
-        TrackingPointMethodsBase
-          .lastExpressionInBlock(block)
-          .map { toTrackedBaseAndAccessPathInternal }
-          .getOrElse((TrackedUnknown, Nil))
-      case call: nodes.Call if !MemberAccess.isGenericMemberAccessName(call.name) => (TrackedReturnValue(call), Nil)
-
-      case memberAccess: nodes.Call =>
-        //assume: MemberAccess.isGenericMemberAccessName(call.name)
-        //FIXME: elevate debug to warn once csharp2cpg has managed to migrate.
-        val argOne = memberAccess.argumentOption(1)
-        if (argOne.isEmpty) {
-          logger.warn(s"Missing first argument on call ${memberAccess}.")
-          return (TrackedUnknown, Nil)
-        }
-        val (base, tail) = toTrackedBaseAndAccessPathInternal(argOne.get)
-        val path = memberAccess.name match {
-          case Operators.memberAccess | Operators.indirectMemberAccess =>
-            if (!hasWarnedDeprecations) {
-              logger.info(s"Deprecated Operator ${memberAccess.name} on ${memberAccess}")
-              hasWarnedDeprecations = true
-            }
-            memberAccess
-              .argumentOption(2)
-              .collect {
-                case lit: nodes.Literal      => ConstantAccess(lit.code)
-                case withName: nodes.HasName => ConstantAccess(withName.name)
-              }
-              .getOrElse(VariableAccess) :: tail
-
-          case Operators.computedMemberAccess | Operators.indirectComputedMemberAccess =>
-            if (!hasWarnedDeprecations) {
-              logger.info(s"Deprecated Operator ${memberAccess.name} on ${memberAccess}")
-              hasWarnedDeprecations = true
-            }
-            memberAccess
-              .argumentOption(2)
-              .collect {
-                case lit: nodes.Literal => ConstantAccess(lit.code)
-              }
-              .getOrElse(VariableAccess) :: tail
-          case Operators.indirection =>
-            IndirectionAccess :: tail
-          case Operators.addressOf =>
-            AddressOf :: tail
-          case Operators.fieldAccess | Operators.indexAccess =>
-            extractAccessStringToken(memberAccess) :: tail
-          case Operators.indirectFieldAccess =>
-            // we will reverse the list in the end
-            extractAccessStringToken(memberAccess) :: IndirectionAccess :: tail
-          case Operators.indirectIndexAccess =>
-            // we will reverse the list in the end
-            IndirectionAccess :: extractAccessIntToken(memberAccess) :: tail
-          case Operators.pointerShift =>
-            extractAccessIntToken(memberAccess) :: tail
-          case Operators.getElementPtr =>
-            // we will reverse the list in the end
-            AddressOf :: extractAccessStringToken(memberAccess) :: IndirectionAccess :: tail
-        }
-        (base, path)
-    }
-
-  private def extractAccessStringToken(memberAccess: nodes.Call): AccessElement = {
-    memberAccess.argumentOption(2) match {
-      case None => {
-        logger.warn(
-          s"Invalid AST: Found member access without second argument." +
-            s" Member access CODE: ${memberAccess.code}" +
-            s" In method ${memberAccess.method.fullName}")
-        VariableAccess
-      }
-      case Some(literal: nodes.Literal) => ConstantAccess(literal.code)
-      case Some(fieldIdentifier: nodes.FieldIdentifier) =>
-        ConstantAccess(fieldIdentifier.canonicalName)
-      case _ => VariableAccess
-    }
-  }
-  private def extractAccessIntToken(memberAccess: nodes.Call): AccessElement = {
-    memberAccess.argumentOption(2) match {
-      case None => {
-        logger.warn(
-          s"Invalid AST: Found member access without second argument." +
-            s" Member access CODE: ${memberAccess.code}" +
-            s" In method ${memberAccess.method.fullName}")
-        VariablePointerShift
-      }
-      case Some(literal: nodes.Literal) =>
-        literal.code.toIntOption.map(PointerShift).getOrElse(VariablePointerShift)
-      case Some(fieldIdentifier: nodes.FieldIdentifier) =>
-        fieldIdentifier.canonicalName.toIntOption
-          .map(PointerShift)
-          .getOrElse(VariablePointerShift)
-      case _ => VariablePointerShift
+      case cfgNode: nodes.CfgNode =>
+        cfgNode.trackedBaseAccessPath({ x =>
+          toTrackedBaseAndAccessPathInternal(x.asInstanceOf[nodes.TrackingPoint])
+        })
     }
   }
